Better ManyToMany relation in database, simple model mapping in conversationservice
