+++ src/main/docker/docker-compose.yml
@@ -31,6 +31,7 @@ services:
       - '3000:3000'
     depends_on:
       - springboot-api-demo
+    stdin_open: true
