+++ src/main/java/com/example/licencjat/conversations/ConversationRepository.java
@@ -1,0 +1,11 @@
+package com.example.licencjat.conversations;
+
+import com.example.licencjat.conversations.models.Conversation;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface ConversationRepository extends JpaRepository<Conversation, UUID> {
+}
