--- src/main/java/com/example/licencjat/conversations/ConversationsCRUD/ConversationsService.java
@@ -1,13 +1,0 @@
-package com.example.licencjat.conversations.ConversationsCRUD;
-
-import com.example.licencjat.conversations.ConversationsCRUD.models.ConversationListDTO;
-import com.example.licencjat.conversations.ConversationsCRUD.models.ConversationWebInput;
-
-import java.util.List;
-import java.util.UUID;
-
-public interface ConversationsService {
-    List<ConversationListDTO> getConversations();
-    void deleteConversation(UUID id);
-    ConversationListDTO createConversation(ConversationWebInput webInput);
-}
