+++ src/main/java/com/example/licencjat/conversations/models/ConversationWebInput.java
@@ -1,0 +1,17 @@
+package com.example.licencjat.conversations.models;
+
+import lombok.Getter;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import java.util.List;
+import java.util.UUID;
+
+@Getter
+public class ConversationWebInput {
+    @NotBlank
+    private String name;
+
+    @NotNull
+    private List<UUID> users;
+}
