+++ src/main/java/com/example/licencjat/messages/models/Message.java
@@ -1,6 +1,6 @@
 package com.example.licencjat.messages.models;
 
-import com.example.licencjat.conversations.ConversationsCRUD.models.Conversation;
+import com.example.licencjat.conversations.models.Conversation;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,6 +8,7 @@ import lombok.Setter;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import java.sql.Timestamp;
 import java.util.UUID;
@@ -20,6 +21,7 @@ public class Message {
     private UUID id;
 
     @ManyToOne
+    @JoinColumn(name = "conversation_id", nullable = false)
     private Conversation conversation;
 
     private UUID ownerId;
