+++ src/main/java/com/example/licencjat/messages/models/MessageOutPut.java
@@ -1,8 +1,6 @@
 package com.example.licencjat.messages.models;
 
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 
 import java.sql.Timestamp;
 import java.util.UUID;
@@ -10,9 +8,10 @@ import java.util.UUID;
 @Getter
 @Setter
 @Builder
+@NoArgsConstructor @AllArgsConstructor
 public class MessageOutPut {
-    private String text;
-    private UUID senderId;
+    private String content;
+    private UUID ownerId;
     private UUID conversationId;
     private Timestamp createdAt;
 }
