+++ src/main/java/com/example/licencjat/user/authorities/models/AuthorityGroup.java
@@ -24,8 +24,4 @@ public class AuthorityGroup {
 
     @ManyToMany(mappedBy = "userGroups")
     private List<User> users = new ArrayList<>();
-
-    public void addUser(User u) {
-        users.add(u);
-    }
 }
